New subprocess implementation

+++ app.py
@@ -15,7 +15,8 @@ p = None
 @app.route('/arrange/<section>', methods=['GET'])
 def home(section):
     global p
-    p.kill()
+    if p is not None:
+        p.kill()
     p = subprocess.Popen("exec python " + LIGHTS_HOME + "child_process.py " + section, shell=True)
     return "<h1>Indoor Lights!</h1><p>"+str(section)+"</p>"
